OC-5438 Create validations for upload of rules where item is not supported for EventAction
AC1: If target is an item, EventAction must not be allowed to be saved as a rule and a meaningful error message must be presented to the user.

AC2: IF the target is an event property of STARTDATE or STATUS then the only supported action must be EventAction and if the user attempts to validate with another action then meaningful error message must be presented to the user.