bugfix, fail to upload the same file during double data entry
Upload a file to the CRF in initial data entry and mark crf complete,
now enter data in double data entry and upload the same file.

There will be error saying "The value you specify does not match the
value from initial data entry".

Problem was caused because the usage of timestamp to ensure file
uniqueness, where using the hash value of the file content is probably
the better option.